implement some Direct2D backend methods

+++ include/impl/RkDirect2DGraphicsBackend.h
@@ -51,7 +51,7 @@ class RkDirect2DGraphicsBackend final : public RkGraphicsBackend {
  private:
         ID2D1HwndRenderTarget *renderTarget;
         ID2D1SolidColorBrush *targetBrush;
-        int strokeWidth;
+        FLOAT strokeWidth;
         ID2D1StrokeStyle *strokeStyle;
 };
